Make PowerScale configuration parameter an int, not a double

+++ begrun.c
@@ -1131,7 +1131,7 @@ void read_parameter_file(char *fname)
 
         strcpy(tag[nt], "DMB_InteractionPowerScale");
         addr[nt] = &All.DMB_InteractionPowerScale;
-        id[nt++] = REAL;
+        id[nt++] = INT;
 #endif
